refactore: database seeds
