+++ backend/src/database/knexfile.ts
@@ -10,6 +10,7 @@ const commonConfig = {
     tableName: 'knex_migrations',
     directory: [`${__dirname}/migrations`],
   },
+  useNullAsDefault: true,
 };
 
 const config: { [key: string]: Knex.Config } = {
