+++ backend/src/database/seeds/mensa-menu.ts
@@ -1,0 +1,5 @@
+import { RepoScheduledTasks } from '../../services';
+
+export async function seed(): Promise<void> {
+  await RepoScheduledTasks.saveMensaMenusToDatabase();
+}
